Add config for minumum light level.
While developing a pack with very large caves, I found that torches can't penetrate the absolute darkness reasonably enough, but having configurable deep darkness and moon phase variations is still desired.
This pull request adds a config for minimum light level. It defaults to 0.0, and therefore, does not change the default behavior of the mod.